Moved everything to settings
